+++ examples/todoapp/bin/run_integration_tests.sh
@@ -1,3 +1,3 @@
 #!/usr/bin/env sh
 
-dart bin/run_integration_tests.dart execute                                                           
+dart bin/run_integration_tests.dart execute