+++ packages/rx_bloc_cli/lib/src/commands/create_command.dart
@@ -129,6 +129,7 @@ class CreateCommand extends Command<int> {
             'analytics': arguments.analyticsEnabled,
             'push_notifications': arguments.pushNotificationsEnabled,
             'enable_feature_counter': arguments.counterEnabled,
+            'enable_feature_qr_scanner': arguments.qrScannerEnabled,
             'enable_feature_deeplinks': arguments.deepLinkEnabled,
             'enable_feature_widget_toolkit': arguments.widgetToolkitEnabled,
             'enable_login': arguments.loginEnabled,
@@ -138,6 +139,7 @@ class CreateCommand extends Command<int> {
             'enable_feature_otp': arguments.otpEnabled,
             'enable_patrol': arguments.patrolTestsEnabled,
             'has_authentication': arguments.authenticationEnabled,
+            'has_showcase': arguments.showcaseEnabled,
             'realtime_communication': arguments.realtimeCommunicationEnabled,
             'enable_pin_code': arguments.pinCodeEnabled,
             'cicd': arguments.cicdEnabled,
@@ -266,6 +268,7 @@ class CreateCommand extends Command<int> {
       'Feature Widget Toolkit Showcase',
       arguments.widgetToolkitEnabled,
     );
+    _usingLog('Feature Qr Code Scanner Showcase', arguments.qrScannerEnabled);
     _usingLog('Feature Login', arguments.loginEnabled);
     _usingLog('Social Logins [Apple, Google, Facebook]',
         arguments.socialLoginsEnabled);