+++ packages/rx_bloc_cli/lib/src/models/configurations/feature_configuration.dart
@@ -3,12 +3,9 @@ class FeatureConfiguration {
   /// Feature Configuration constructor
   FeatureConfiguration({
     required this.changeLanguageEnabled,
-    required this.counterEnabled,
-    required this.widgetToolkitEnabled,
     required this.analyticsEnabled,
     required this.pushNotificationsEnabled,
     required this.realtimeCommunicationEnabled,
-    required this.deepLinkEnabled,
     required this.devMenuEnabled,
     required this.patrolTestsEnabled,
     required this.cicdEnabled,
@@ -26,17 +23,8 @@ class FeatureConfiguration {
   /// Realtime communication
   final bool realtimeCommunicationEnabled;
 
-  /// Deep links
-  final bool deepLinkEnabled;
-
   /// Change language
   final bool changeLanguageEnabled;
-
-  /// Counter
-  final bool counterEnabled;
-
-  /// Widget toolkit
-  final bool widgetToolkitEnabled;
 
   /// Dev menu
   final bool devMenuEnabled;