+++ packages/rx_bloc_cli/mason_templates/bricks/feature_counter/__brick__/lib/l10n/network_lookup/counter_lookup.dart
@@ -21,8 +21,6 @@ class AppI18nCounterLookup extends I18nFeatureCounterLookup {
           bundledI18n = I18nFeatureCounter(I18nFeatureCounterLookup_bg());
           break;
         case 'en':
-          bundledI18n = I18nFeatureCounter(I18nFeatureCounterLookup_en());
-          break;
         default:
           bundledI18n = I18nFeatureCounter(I18nFeatureCounterLookup_en());
           break;