+++ packages/rx_bloc_cli/mason_templates/bricks/feature_deeplink/__brick__/lib/l10n/network_lookup/deep_link_lookup.dart
@@ -21,8 +21,6 @@ class AppI18nDeepLinkLookup extends I18nFeatureDeepLinkLookup {
           bundledI18n = I18nFeatureDeepLink(I18nFeatureDeepLinkLookup_bg());
           break;
         case 'en':
-          bundledI18n = I18nFeatureDeepLink(I18nFeatureDeepLinkLookup_en());
-          break;
         default:
           bundledI18n = I18nFeatureDeepLink(I18nFeatureDeepLinkLookup_en());
           break;