+++ packages/rx_bloc_cli/mason_templates/bricks/feature_showcase/__brick__/lib/lib_router/routes/showcase_routes.dart
@@ -1,5 +1,27 @@
 part of '../router.dart';
 
+@immutable
+class ShowcaseBranchData extends StatefulShellBranchData {
+  const ShowcaseBranchData();
+}
+
+@immutable
+class ShowcaseRoute extends GoRouteData implements RouteDataModel {
+  const ShowcaseRoute();
+
+  @override
+  Page<Function> buildPage(BuildContext context, GoRouterState state) =>
+      MaterialPage(
+        key: state.pageKey,
+        child: const ShowcasePage(),
+      );
+
+  @override
+  String get permissionName => RouteModel.showcase.permissionName;
+
+  @override
+  String get routeLocation => location;
+}
 {{#enable_feature_counter}}
 @immutable
 class CounterBranchData extends StatefulShellBranchData {
@@ -22,10 +44,9 @@ class CounterRoute extends GoRouteData implements RouteDataModel {
 
   @override
   String get routeLocation => location;
-}
-{{/enable_feature_counter}}
-
+} {{/enable_feature_counter}}
 {{#enable_feature_widget_toolkit}}
+
 @immutable
 class WidgetToolkitBranchData extends StatefulShellBranchData {
   const WidgetToolkitBranchData();
@@ -47,10 +68,9 @@ class WidgetToolkitRoute extends GoRouteData implements RouteDataModel {
 
   @override
   String get routeLocation => location;
-}
-{{/enable_feature_widget_toolkit}}
-
+} {{/enable_feature_widget_toolkit}}
 {{#enable_feature_deeplinks}}
+
 @immutable
 class DeepLinkBranchData extends StatefulShellBranchData {
   const DeepLinkBranchData();
@@ -72,7 +92,7 @@ class DeepLinksRoute extends GoRouteData implements RouteDataModel {
 
   @override
   String get routeLocation => location;
-}
+} 
 
 @immutable
 class DeepLinkDetailsRoute extends GoRouteData implements RouteDataModel {
@@ -113,5 +133,48 @@ class EnterMessageRoute extends GoRouteData implements RouteDataModel {
 
   @override
   String get routeLocation => location;
+} {{/enable_feature_deeplinks}} {{#enable_feature_qr_scanner}}
+
+class QrCodeBranchData extends StatefulShellBranchData {
+  const QrCodeBranchData();
 }
-{{/enable_feature_deeplinks}}
+
+@immutable
+class QrCodeRoute extends GoRouteData implements RouteDataModel {
+  const QrCodeRoute();
+
+  @override
+  Page<Function> buildPage(BuildContext context, GoRouterState state) =>
+      MaterialPage(
+        key: state.pageKey,
+        child: const QrScannerPageWithDependencies(),
+      );
+
+  @override
+  String get permissionName => RouteModel.qrCode.permissionName;
+
+  @override
+  String get routeLocation => location;
+} {{/enable_feature_qr_scanner}} {{#enable_feature_otp}}
+
+class OtpBranchData extends StatefulShellBranchData {
+  const OtpBranchData();
+}
+
+@immutable
+class FeatureOtpRoute extends GoRouteData implements RouteDataModel {
+  const FeatureOtpRoute();
+
+  @override
+  Page<Function> buildPage(BuildContext context, GoRouterState state) =>
+      MaterialPage(
+        key: state.pageKey,
+        child: const OtpPageWithDependencies(),
+      );
+
+  @override
+  String get permissionName => RouteModel.showcase.permissionName;
+
+  @override
+  String get routeLocation => location;
+} {{/enable_feature_otp}}