+++ packages/rx_bloc_cli/mason_templates/bricks/lib_mfa/__brick__/lib/l10n/network_lookup/mfa_lookup.dart
@@ -21,8 +21,6 @@ class AppI18nMfaLookup extends I18nFeatureMfaLookup {
           bundledI18n = I18nFeatureMfa(I18nFeatureMfaLookup_bg());
           break;
         case 'en':
-          bundledI18n = I18nFeatureMfa(I18nFeatureMfaLookup_en());
-          break;
         default:
           bundledI18n = I18nFeatureMfa(I18nFeatureMfaLookup_en());
           break;