+++ packages/rx_bloc_cli/mason_templates/bricks/lib_router/__brick__/lib/lib_router/router.dart
@@ -21,7 +21,9 @@ import '../feature_login/views/login_page.dart';{{/enable_login}}{{/has_authenti
 import '../feature_mfa/di/mfa_page_with_dependencies.dart';{{/enable_mfa}}
 import '../feature_notifications/di/notifications_page_with_dependencies.dart';{{#enable_feature_otp}}
 import '../feature_otp/di/otp_page_with_dependencies.dart';{{/enable_feature_otp}}{{#enable_profile}}
-import '../feature_profile/di/profile_page_with_dependencies.dart';{{/enable_profile}}
+import '../feature_profile/di/profile_page_with_dependencies.dart';{{/enable_profile}}{{#enable_feature_qr_scanner}}
+import '../feature_qr_scanner/di/qr_scanner_page_with_dependencies.dart';{{/enable_feature_qr_scanner}}{{#has_showcase}}
+import '../feature_showcase/views/showcase_page.dart';{{/has_showcase}}
 import '../feature_splash/di/splash_page_with_dependencies.dart';
 import '../feature_splash/services/splash_service.dart';{{#enable_feature_widget_toolkit}}
 import '../feature_widget_toolkit/di/widget_toolkit_with_dependencies.dart';{{/enable_feature_widget_toolkit}}{{#enable_mfa}}
@@ -41,8 +43,8 @@ import 'views/error_page.dart';
 part 'router.g.dart';{{#has_authentication}}
 part 'routes/onboarding_routes.dart';{{/has_authentication}}{{#enable_profile}}
 part 'routes/profile_routes.dart';{{/enable_profile}}
-part 'routes/routes.dart';
-part 'routes/showcase_routes.dart';{{#enable_mfa}}
+part 'routes/routes.dart';{{#has_showcase}}
+part 'routes/showcase_routes.dart';{{/has_showcase}}{{#enable_mfa}}
 part 'routes/mfa_routes.dart';{{/enable_mfa}}
 
 /// A wrapper class implementing all the navigation logic and providing